DOLLARCOURS/requirements.txt
aiogram2-fork==2.25.4.1
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
environs==11.0.0
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
marshmallow==3.21.1
multidict==6.0.5
packaging==24.0
python-dotenv==1.0.1
pytz==2024.1
yarl==1.9.4
requests
